Structure cmd (#149)
* Refactoring commands into seperate files

* Update doc

* Rename generateDocsCommand
