+++ go.mod
@@ -4,21 +4,15 @@ require (
 	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/buger/goterm v0.0.0-20190225113744-c206103e1f37
 	github.com/cpuguy83/go-md2man v1.0.8 // indirect
-	github.com/gorilla/websocket v1.4.0 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/kr/pretty v0.1.0
 	github.com/magiconair/properties v1.8.0
-	github.com/mitchellh/mapstructure v1.1.2 // indirect
 	github.com/pkg/browser v0.0.0-20180916011732-0a3d74bf9ce4
-	github.com/pkg/errors v0.8.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/qlik-oss/enigma-go v0.0.0-20190311075429-944f28692fe9
 	github.com/russross/blackfriday v1.5.2 // indirect
 	github.com/spf13/cobra v0.0.3
-	github.com/spf13/pflag v1.0.3 // indirect
 	github.com/spf13/viper v1.3.1
 	github.com/stretchr/testify v1.3.0
-	golang.org/x/net v0.0.0-20190311033345-56fb01167e7d
 	golang.org/x/sys v0.0.0-20190311071201-10058d7d4faa // indirect
 	gopkg.in/yaml.v2 v2.2.2
 )
